Initial marv test work
Still some things missing, for example building the sim in the makefile. I'm considering also moving the sim to this repo instead of having it as part of the xenowing repo, but I'm not sure yet.

There's also a few hacks; for example, the init code is heavier than it needs to be (probably), and I changed some default script settings to match my local setup.