Place .tohost section before .data section
It seems the SW test (and likely SH/SB as well) writes a word _before_ the target address in the first test. This was caught as a write to an unknown address in the marv test harness.

Looking at the default RVTEST_DATA_BEGIN/END, it seems it's expected that .tohost is before .data, or perhaps not. Either way, this makes the tests pass, and I'm quite OK with that, as I haven't spotted an error in the RTL code for this when double-checking.